Adjusting namespaces
