+++ src/Equinor.ProCoSys.IPO.Domain/Events/PreSave/AttachmentRemovedEvent.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using MediatR;
 
-namespace Equinor.ProCoSys.IPO.Domain.Events
+namespace Equinor.ProCoSys.IPO.Domain.Events.PreSave
 {
     public class AttachmentRemovedEvent : INotification
     {
