+++ src/Equinor.ProCoSys.IPO.WebApi/Seeding/Seeder.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
